Add fields of interest processing
After playing a bit, I figured out SCons does not like running packages.
it doesn't properly track the fact that the files changed so it defeats
the purpose.  So, we begin migrating to a single script so we know it
will get updated if we make an adjustment.  This does mean we will have
to constantly rerun, but the script is fast.