Add proposal information as a folder
To keep things a bit clearer (and to actually be able to find
something), I think it will be better to keep the proposals in one
per file. So far, I'm mainly using the F# for the proposal, the project
number, or just some other unique file name.  So long as it all works.
The only thing we need to do then is merge the YAML files into one large
file under the 'proposals' key for the next script to work.