Add additional key delivered products
The LaTeX products I've worked on are relevant.  The promotion packet
one is a bit of a stretch, but it could point out to higher ups that
Word is not the best way to manage such a complex document.  Or, it
starts getting the word out that someone has put together a "better" way
to manage their CV for promotion.