Update yaml merge script to deal with lists
Some of the proposals, like the Warner Robins white paper, are repeated.
It will be a little bit simpler to just keep those together instead of
figuring out how to deal with separate names.  Now the script will
flatten a list to make one big list to use in the output file.